feat: Refactor PDF parser into modular architecture
