+++ src/__init__.py
@@ -1,0 +1,6 @@
+from .pdfmind.pdf_parser import PDFParser
+from .pdfmind.processor.image_processor import ImageProcessor
+from .pdfmind.processor.table_processor import TableProcessor
+from .pdfmind.processor.toc_processor import TOCProcessor
+
+__all__ = ["PDFParser", "ImageProcessor", "TableProcessor", "TOCProcessor"]
