+++ src/pdfmind/processor/image_processor.py
@@ -1,0 +1,17 @@
+import logging
+from typing import List, Dict, Any
+
+
+class ImageProcessor:
+    """Handles extraction and processing of images from PDF documents.
+    
+    This module is designed for future implementation to handle:
+    - Image extraction from PDF files
+    - Image format conversion and optimization
+    - Image metadata extraction
+    - Image classification and analysis
+    """
+    
+    def __init__(self):
+        """Initialize the image processor."""
+        logging.info("ImageProcessor initialized (placeholder implementation)")
