+++ src/pdfmind/prompts/__init__.py
@@ -1,0 +1,3 @@
+from .prompts import TableOfContentsExtractor
+
+__all__ = ['TableOfContentsExtractor']
