+++ src/pdfmind/prompts/prompts.py
